Move SMN to central location.
