+++ src/mainboard/amd/romecrb/src/main.rs
@@ -13,8 +13,8 @@ use cpu::model::amd_model_id;
 use model::Driver;
 use print;
 use raw_cpuid::CpuId;
+use smn::{smn_read, smn_write};
 use soc::soc_init;
-use soc::{smn_read, smn_write};
 use uart::amdmmio::AMDMMIO;
 use uart::debug_port::DebugPort;
 use uart::i8250::I8250;
