+++ src/mainboard/amd/romecrb/src/mainboard.rs
@@ -22,6 +22,7 @@ use core::ops::BitOr;
 use core::ops::Not;
 use core::ptr;
 use model::*;
+use smn::smn_write;
 use vcell::VolatileCell;
 use x86_64::registers::model_specific::Msr;
 
@@ -62,36 +63,6 @@ const FCH_AOAC_DEV_UART0: u8 = 11;
 const FCH_AOAC_DEV_UART1: u8 = 12;
 const FCH_AOAC_DEV_AMBA: u8 = 17;
 
-// It's kind of a shame, but every single pci crate I've looked at is
-// basically close to useless. Unless I'm missing something,
-// which is likely. They really should get all the various authors
-// and a room and just DEFINE ONE THING. It's not rocket science.
-// I'm not going to attempt to write one because:
-// 1. I suck at it.
-// 2. It would be JUST ONE MORE.
-// SMN
-
-fn smn_read(a: u32) -> u32 {
-    // the smn device is at (0)
-    unsafe {
-        outl(0xcf8, 0x8000_00b8);
-        outl(0xcfc, a);
-        outl(0xcf8, 0x8000_00bc);
-        inl(0xcfc)
-    }
-}
-
-fn smn_write(a: u32, v: u32) {
-    unsafe {
-        outl(0xcf8, 0x800000b8);
-        outl(0xcfc, a);
-        outl(0xcf8, 0x800000bc);
-        outl(0xcfc, v);
-    }
-}
-
-// end SMN
-
 unsafe fn poke<T>(a: *mut T, v: T) -> () {
     ptr::write_volatile(a, v);
 }
@@ -106,18 +77,6 @@ where
     T: Copy + Not<Output = T> + BitAnd<Output = T> + BitOr<Output = T>,
 {
     (*a).set(((*a).get() & !reset_mask) | set_mask);
-}
-
-/// Write 32 bits to port
-unsafe fn outl(port: u16, val: u32) {
-    llvm_asm!("outl %eax, %dx" :: "{dx}"(port), "{al}"(val));
-}
-
-/// Read 32 bits from port
-unsafe fn inl(port: u16) -> u32 {
-    let ret: u32;
-    llvm_asm!("inl %dx, %eax" : "={ax}"(ret) : "{dx}"(port) :: "volatile");
-    return ret;
 }
 
 // WIP: mainboard driver. I mean the concept is a WIP.
