+++ src/soc/amd/common/smn/Cargo.toml
@@ -1,14 +1,11 @@
 [package]
-name = "soc"
+name = "smn"
 version = "0.1.0"
 authors = ["Oreboot Authors"]
 edition = "2018"
 
 [dependencies]
-arch = { path = "../../../arch/x86/x86_64" }
-df = { path = "../common/df" }
-pci = { path = "../common/pci" }
-smn = { path = "../common/smn" }
+arch = { path = "../../../../arch/x86/x86_64"}
 vcell = "0.1.2"
 
 [profile.release]
